Make menuAddScreenEdgePanGesturesToPresent compatible with Obj-C
Swift has optionals for all types which works in most cases when being
used in Obj-C projects. However optional primitive types
(non-heap-allocated, like enums in this case) can’t be converted, and
whatever variables of functions that use them are invisible to Obj-C.

Another impossibility to convert was the default parameter, so I had to
add convenience function to call the full one.